+++ src/hi/integrations/connect/external_view_data.py
