+++ src/hi/integrations/connect/integration_manage_view_pane.py
