+++ src/hi/integrations/integration_manager.py
@@ -18,11 +18,11 @@ from hi.apps.common.module_utils import import_module_safe
 from hi.apps.entity.models import Entity
 from hi.apps.system.health_status_provider import HealthStatusProvider
 
-from .entity_operations import EntityIntegrationOperations
+from .connect.entity_operations import EntityIntegrationOperations
 from .enums import IntegrationAttributeType, IntegrationDisableMode
 from .exceptions import IntegrationConnectionError
-from .integration_data import IntegrationData
-from .integration_gateway import IntegrationGateway
+from .connect.integration_data import IntegrationData
+from .connect.integration_gateway import IntegrationGateway
 from .transient_models import IntegrationKey
 from .models import Integration, IntegrationAttribute
 from .transient_models import IntegrationMetaData
@@ -251,7 +251,7 @@ class IntegrationManager( Singleton ):
         # monitor module imports IntegrationManager lazily inside its
         # do_work, but the manager only needs the class here for
         # construction.
-        from .monitors import IntegrationSyncCheckMonitor
+        from .connect.monitors import IntegrationSyncCheckMonitor
 
         if settings.DEBUG and settings.SUPPRESS_MONITORS:
             logger.debug( 'Skipping sync-check monitor. See SUPPRESS_MONITORS = True' )
