+++ src/hi/services/homebox/connector/hb_sync.py
@@ -6,9 +6,9 @@ from django.db import transaction
 
 from hi.apps.entity.models import Entity
 
-from hi.integrations.integration_synchronizer import IntegrationSynchronizer
-from hi.integrations.sync_check import IntegrationSyncCheck, SyncDelta
-from hi.integrations.sync_result import IntegrationSyncResult
+from hi.integrations.connect.integration_synchronizer import IntegrationSynchronizer
+from hi.integrations.connect.sync_check import IntegrationSyncCheck, SyncDelta
+from hi.integrations.connect.sync_result import IntegrationSyncResult
 from hi.integrations.transient_models import IntegrationKey
 
 from hi.services.homebox.shared.hb_converter import HbConverter
