+++ src/hi/services/homebox/importer/hb_attribute_sync.py
@@ -9,7 +9,7 @@ from django.db import transaction
 
 from hi.apps.entity.models import Entity, EntityAttribute
 
-from hi.integrations.sync_result import IntegrationSyncResult
+from hi.integrations.connect.sync_result import IntegrationSyncResult
 from hi.integrations.transient_models import IntegrationKey
 
 from hi.services.homebox.shared.hb_converter import HbConverter
